Use the free-list with utl::vector and std::vector
Previously, I wrote a free-list container which had a vector as its buffer and I ran into the problem of conflicting destruction of items which made it necessary to write our own vector container that has an extra template parameter to control whether the items should be destructed upon deletion or not. I stopped halfway of writing the vector class and today I'm going to finish it and use it for the free-list.